backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.13.0",
    "bcrypt>=4.0.0",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.40.0",
]
